Task 101 v7

+++ 0x00-hello_world/101-quote.c
@@ -10,7 +10,7 @@ int main(void)
 {
 	char str[] = "and that piece of art is useful\" - Dora Korpar, 2015-10-19";
 
-	putchar(str);
+	putchar((int)str);
 	/* return an int */
 	return (1);
 }
